feat: 로그아웃 구현

+++ src/features/auth/hooks/useAuth.ts
@@ -85,11 +85,15 @@ export const useAuth = () => {
       alert('로그아웃을 다시 시도해주십시요.');
       return;
     }
+    const confirm: boolean = window.confirm('로그아웃 하시겠습니까?');
+    if (!confirm) return;
+
     logoutMutation.mutate(accessToken);
     localStorage.removeItem('accessToken');
     localStorage.removeItem('user');
     dispatch(removeUser());
     dispatch(removeAccessToken());
+    router.push('/');
   };
 
   return { openLoginPopup, logout };
